Fix update
Script comments changed to english.
Removed System.Collections or System.Collections.Generic from script in order for the script to be cleaner and not to create confusion.
Changed RigidBody2D to RigidBody2D ball in script in order to be more meaningfull and not to create confusion.
Removed commets from start() and update() function in script in order for the code to be cleaner.
Created seperate folders for each type of files.